updated multiplier

+++ causal_effect_estimation/code/BBB_ours.py
@@ -23,7 +23,7 @@ def main():
     ''' evaluate true network causal effects '''
     _, network_adj_mat = create_random_network(TRUE_CAUSAL_EFFECT_N_UNIT, AVG_DEGREE, MAX_NEIGHBORS)
     causal_effect_true = true_causal_effects_B_B(network_adj_mat, L_TRUE, Y_TRUE,
-                                                 N_SIM_MULTIPLIER*TRUE_CAUSAL_EFFECT_N_UNIT)
+                                                 int(N_SIM_MULTIPLIER*TRUE_CAUSAL_EFFECT_N_UNIT))
     print("True causal effect:", causal_effect_true)
     
     ''' using autog to estimate causal effects from data generated from BBB '''
